+++ ical-recurrence/default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "ical-recurrence";
-  version = "0.0.0.0";
+  version = "0.1.0.0";
   src = ./.;
   libraryHaskellDepends = [
     base conformance containers ical mtl time validity